Project 5 - Week 11
Making media queries that changes the styling to incorporate three media
types (desktop, tablet, and phone).
Changed the navigation so the images change size when on a tablet (so
the desktop images aren't too big) and then removed the graphics, making
the navigation just text (also removing the sub-menus) for a phone.
I changed the size of the logo for phones, as the desktop version is too
big.
I changed the padding and the font size for the smaller devices also.